Rename bash_test to bash_tests
